12-Client-Deliverables/05-Decision-Logs/requirements_http_mcp.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.6.1
supabase==2.3.0
python-dotenv==1.0.0
